+++ Tests/Tracking/Util/SimpleTimerTest.swift
@@ -3,25 +3,25 @@ import Foundation
 import SharedTests
 import XCTest
 
-class SingleScheduleTimerTest: UnitTest {
-    private var timer: SingleScheduleTimer!
+class SimpleTimerTest: UnitTest {
+    private var timer: CioSimpleTimer!
 
     override func setUp() {
         super.setUp()
 
-        timer = CioSingleScheduleTimer(timer: diGraph.simpleTimer)
+        timer = CioSimpleTimer(logger: log)
     }
 
-    // MARK: scheduleIfNotAleady
+    // MARK: scheduleIfNotAlready
 
-    func test_scheduleIfNotAleady_givenCallMultipleTimes_expectIgnoreFutureRequests() {
+    func test_scheduleIfNotAlready_givenCallMultipleTimes_expectIgnoreFutureRequests() {
         let expect = expectation(description: "Timer fires")
-        var didSchedule = timer.scheduleIfNotAleady(numSeconds: 0.1) {
+        var didSchedule = timer.scheduleIfNotAlready(seconds: 0.1) {
             expect.fulfill()
         }
         XCTAssertTrue(didSchedule)
 
-        didSchedule = timer.scheduleIfNotAleady(numSeconds: 0) {
+        didSchedule = timer.scheduleIfNotAlready(seconds: 0) {
             expect.fulfill() // this should not fire
         }
         XCTAssertFalse(didSchedule)
@@ -29,19 +29,38 @@ class SingleScheduleTimerTest: UnitTest {
         waitForExpectations()
     }
 
-    func test_scheduleIfNotAleady_givenCallAfterTimerFires_expectStartNewTimer() {
+    func test_scheduleIfNotAlready_givenCallAfterTimerFires_expectStartNewTimer() {
         var expect = expectation(description: "Timer fires")
-        var didSchedule = timer.scheduleIfNotAleady(numSeconds: 0) {
+        var didSchedule = timer.scheduleIfNotAlready(seconds: 0) {
             expect.fulfill()
         }
         XCTAssertTrue(didSchedule)
         waitForExpectations()
 
         expect = expectation(description: "Timer fires")
-        didSchedule = timer.scheduleIfNotAleady(numSeconds: 0) {
+        didSchedule = timer.scheduleIfNotAlready(seconds: 0) {
             expect.fulfill()
         }
         XCTAssertTrue(didSchedule)
+        waitForExpectations()
+    }
+
+    // MARK: scheduleAndCancelPrevious
+
+    func test_schedule_givenPreviouslyRunningTimer_expectCancelAndStartNew() {
+        let expectNoFire = expectation(description: "Timer does not fire")
+        expectNoFire.isInverted = true
+        timer.scheduleAndCancelPrevious(seconds: 1) {
+            expectNoFire.fulfill()
+        }
+
+        timer.cancel()
+
+        let expectFire = expectation(description: "Timer fires")
+        timer.scheduleAndCancelPrevious(seconds: 0.1) {
+            expectFire.fulfill()
+        }
+
         waitForExpectations()
     }
 
@@ -54,7 +73,7 @@ class SingleScheduleTimerTest: UnitTest {
     func test_cancel_givenScheduled_expectTimerCanceled() {
         let expect = expectation(description: "Timer does not fire")
         expect.isInverted = true
-        let didSchedule = timer.scheduleIfNotAleady(numSeconds: 0.1) {
+        let didSchedule = timer.scheduleIfNotAlready(seconds: 0.1) {
             expect.fulfill()
         }
         XCTAssertTrue(didSchedule)
@@ -67,7 +86,7 @@ class SingleScheduleTimerTest: UnitTest {
     func test_cancel_expectScheduleAfterCancel() {
         let expectNoFire = expectation(description: "Timer does not fire")
         expectNoFire.isInverted = true
-        var didSchedule = timer.scheduleIfNotAleady(numSeconds: 0.1) {
+        var didSchedule = timer.scheduleIfNotAlready(seconds: 0.1) {
             expectNoFire.fulfill()
         }
         XCTAssertTrue(didSchedule)
@@ -75,7 +94,7 @@ class SingleScheduleTimerTest: UnitTest {
         timer.cancel()
 
         let expectFire = expectation(description: "Timer fires")
-        didSchedule = timer.scheduleIfNotAleady(numSeconds: 0.1) {
+        didSchedule = timer.scheduleIfNotAlready(seconds: 0.1) {
             expectFire.fulfill()
         }
         XCTAssertTrue(didSchedule)